rewrite of code base with pxi design system
this is a bit drastic, but going forward we hope using a design system will help both our development and for others to pick up the code. this is not a perfect transition, some of the javascript logic was tied to proprietary design rules that fall outside of the design system. the system itself is also still evolving.